+++ mobile/android/android-components/components/browser/session/src/test/java/mozilla/components/browser/session/engine/EngineObserverTest.kt
@@ -27,7 +27,6 @@ import mozilla.components.support.test.mock
 import mozilla.components.support.test.whenever
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
-import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
 import org.junit.Test
@@ -525,34 +524,6 @@ class EngineObserverTest {
 
         observer.onCrash()
         assertTrue(session.crashed)
-    }
-
-    @Test
-    fun `onLocationChange clears icon`() {
-        val session = Session("https://www.mozilla.org")
-        session.icon = mock()
-
-        val observer = EngineObserver(session)
-
-        assertNotNull(session.icon)
-
-        observer.onLocationChange("https://www.firefox.com")
-
-        assertNull(session.icon)
-    }
-
-    @Test
-    fun `onLocationChange doesn't clear icon when new URL is from the same host as the session URL`() {
-        val session = Session("https://www.mozilla.org/?desktop-redirect=true")
-        session.icon = mock()
-
-        val observer = EngineObserver(session)
-
-        assertNotNull(session.icon)
-
-        observer.onLocationChange("https://www.mozilla.org")
-
-        assertNotNull(session.icon)
     }
 
     @Test